client: don't use special inode for /..
The CEPH_INO_DOTDOT thing is quite strange. Under most OS (Linux
included), the parent of the root is itself. IOW, at the root, '.' and
'..' refer to the same inode.

Change the ceph client to do the same, as this allows users to get
valid stat info for '..', as well as elimnating some special-casing.

Also in several places, we're checking dn_set.empty as an indicator
of being the root. While that is true for the root, it's also true
for unlinked directories.

We have an Inode->is_root() method that is currently unused and
incorrect. What we really need is a way to see whether the inode is the
root of its Client.  Remove the Inode->is_root() and new methods to the
Client to tell us that.  Then, use that in place of the dn_set.empty
test in the handling of "..".

Signed-off-by: Jeff Layton <jlayton@redhat.com>